+++ src/pyubx2/ubxtypes_core.py
@@ -12,6 +12,7 @@ UBX_HDR = b"\xb5\x62"
 GET = 0
 SET = 1
 POLL = 2
+SETPOLL = 3
 VALNONE = 0
 VALCKSUM = 1
 NMEA_PROTOCOL = 1
@@ -20,17 +21,6 @@ RTCM3_PROTOCOL = 4
 ERR_RAISE = 2
 ERR_LOG = 1
 ERR_IGNORE = 0
-
-GNSSLIST = {
-    0: "GPS",
-    1: "SBAS",
-    2: "Galileo",
-    3: "BeiDou",
-    4: "IMES",
-    5: "QZSS",
-    6: "GLONASS",
-    7: "NAVIC",
-}
 
 # scaling factor constants
 SCAL9 = 1e-9  # 0.000000001